+++ tests/support/test_supported.py
@@ -6,7 +6,7 @@ class TestCase(unittest.TestCase):
 
     def test_import_and_context(self):
         ModernGL = importlib.import_module('ModernGL')
-        ctx = ModernGL.create_standalone_context()
+        ctx = ModernGL.create_standalone_context(require=ModernGL.CORE_330)
         ctx.release()