+++ tests/test_vertex_attribute_types.py
@@ -5,205 +5,205 @@ import ModernGL
 
 vtypes = [
     {
-        'version': '330',
+        'version': 330,
         'type': 'int',
         'input': [-1],
         'output': [-2],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'ivec2',
         'input': [-1, -1],
         'output': [-2, -2],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'ivec3',
         'input': [-1, -1, -1],
         'output': [-2, -2, -2],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'ivec4',
         'input': [-1, -1, -1, -1],
         'output': [-2, -2, -2, -2],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'uint',
         'input': [1],
         'output': [2],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'uvec2',
         'input': [1, 1],
         'output': [2, 2],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'uvec3',
         'input': [1, 1, 1],
         'output': [2, 2, 2],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'uvec4',
         'input': [1, 1, 1, 1],
         'output': [2, 2, 2, 2],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'float',
         'input': [1.0],
         'output': [2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'vec2',
         'input': [1.0, 1.0],
         'output': [2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'vec3',
         'input': [1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'vec4',
         'input': [1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'double',
         'input': [1.0],
         'output': [2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dvec2',
         'input': [1.0, 1.0],
         'output': [2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dvec3',
         'input': [1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dvec4',
         'input': [1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'mat2',
         'input': [1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'mat2x3',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'mat2x4',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'mat3x2',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'mat3',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'mat3x4',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'mat4x2',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'mat4x3',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '330',
+        'version': 330,
         'type': 'mat4',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dmat2',
         'input': [1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dmat2x3',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dmat2x4',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dmat3x2',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dmat3',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dmat3x4',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dmat4x2',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dmat4x3',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
     },
     {
-        'version': '410',
+        'version': 410,
         'type': 'dmat4',
         'input': [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
         'output': [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0],
@@ -229,16 +229,15 @@ class TestCase(unittest.TestCase):
             out %(type)s v_out;
 
             void main() {
-                v_out = v_in * 2;
+                v_out = v_in + v_in;
             }
         '''
 
         for vtype in vtypes:
-            try:
-                prog = self.ctx.program(self.ctx.vertex_shader(vert_src % vtype), ['v_out'])
-            except ModernGL.Error:
-                # skip when version 410 not supported
+            if self.ctx.version_code < vtype['version']:
                 continue
+
+            prog = self.ctx.program(self.ctx.vertex_shader(vert_src % vtype), ['v_out'])
 
             fmt = ModernGL.detect_format(prog, ['v_in'])
             vbo1 = self.ctx.buffer(struct.pack(fmt, *vtype['input']))
@@ -257,17 +256,16 @@ class TestCase(unittest.TestCase):
             out %(type)s v_out[];
 
             void main() {
-                v_out[0] = v_in[0] * 2;
-                v_out[1] = v_in[1] * 2;
+                v_out[0] = v_in[0] + v_in[0];
+                v_out[1] = v_in[1] + v_in[1];
             }
         '''
 
         for vtype in vtypes:
-            try:
-                prog = self.ctx.program(self.ctx.vertex_shader(vert_src % vtype), ['v_out'])
-            except ModernGL.Error:
-                # skip when version 410 not supported
+            if self.ctx.version_code < vtype['version']:
                 continue
+
+            prog = self.ctx.program(self.ctx.vertex_shader(vert_src % vtype), ['v_out'])
 
             fmt = ModernGL.detect_format(prog, ['v_in'])
             vbo1 = self.ctx.buffer(struct.pack(fmt, *(vtype['input'] * 2)))